(SLV-774) Restart compilers in scale_setup
This commit updates the call to `restart_puppet_server` in
`scale_setup` to iterate over all hosts with the roles of `master` and
`compile_master`.  Prior to this commit only the master was restarted
which is not sufficient in a deployment that uses compile masters in
addition to the core master server.