(SLV-769) performance_provision_with_abs add large
This commit updates the `performance_provision_with_abs` rake task to
provision a set of hosts for a PE large architecture if the `REF_ARCH`
environment variable is set to large.

Environment variables for setting AWS size and volume size are supported
for all of the roles within the large architecture.

Since the `get_a2a_hosts` helper is no longer used, it has been deleted
along with the abs_helper instance variables that were only used within
it.