(maint) Update 2018.1 a2a variances
This commit updates the variances for the following performance metrics:

* process_orchestration_services_release_avg_mem
* avg_disk_write
* puppetdb_avg_mem

Prior to this commit, the PE master branch builds consistently failed
these metrics when compared to to the 2018.1.x LTS version.  These
changes update the variances to fall above the highest failure point.
Since the 2018.1.x LTS will be replaced soon, engineering effort to
invesigate these metric failures will not be applied.

Ref:
[PE-28214](https://tickets.puppetlabs.com/browse/PE-28214)
[PE-28215](https://tickets.puppetlabs.com/browse/PE-28215)
[PDB-4635](https://tickets.puppetlabs.com/browse/PDB-4635)